refactor: remove 'da' suffix and rename to `client.go`
