+++ avail/verify/bindings/VectorVerifier/VectorVerifier.go
