+++ avail/verify/contracts/.github/workflows/test.yml
