+++ avail/verify/contracts/src/VectorVerifier.sol
