+++ avail/verify/contracts/src/interfaces/IVectorX.sol
