+++ celestia/verify/bindings/BlobstreamVerifier/BlobstreamVerifier.go
