+++ celestia/verify/contracts/.gitignore
