+++ cmd/blob-server/main.go
@@ -12,9 +12,9 @@ import (
 	"github.com/cenkalti/backoff/v4"
 	"github.com/gin-gonic/gin"
 	"github.com/stackrlabs/go-daash"
-	"github.com/stackrlabs/go-daash/availda"
-	availVerify "github.com/stackrlabs/go-daash/availda/verify"
-	celestiaVerify "github.com/stackrlabs/go-daash/celestiada/verify"
+	"github.com/stackrlabs/go-daash/avail"
+	availVerify "github.com/stackrlabs/go-daash/avail/verify"
+	celestiaVerify "github.com/stackrlabs/go-daash/celestia/verify"
 	"github.com/stackrlabs/go-daash/da"
 )
 
@@ -202,12 +202,12 @@ func verifyDA(c *gin.Context, layer daash.DALayer, daasher *daash.DABuilder) {
 			return
 		}
 		verifier, err := availVerify.NewVerifier(
-			daasher.Clients[daash.Avail].(*availda.Client),
+			daasher.Clients[daash.Avail].(*avail.Client),
 			chainMetadata["sepolia"]["rpcUrl"],
 			chainMetadata["sepolia"]["availBridgeAddress"],
 			chainMetadata["sepolia"]["vectorVerifierAddress"],
 			chainMetadata["sepolia"]["vectorXAddress"],
-			daasher.Clients[daash.Avail].(*availda.Client).Config.Network,
+			daasher.Clients[daash.Avail].(*avail.Client).Config.Network,
 		)
 		if err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{
@@ -215,7 +215,7 @@ func verifyDA(c *gin.Context, layer daash.DALayer, daasher *daash.DABuilder) {
 			})
 			return
 		}
-		success, err = verifier.IsDataIncluded(availda.ID{Height: blockHeightUint, ExtIndex: uint32(extIndexUint)})
+		success, err = verifier.IsDataIncluded(avail.ID{Height: blockHeightUint, ExtIndex: uint32(extIndexUint)})
 		if err != nil {
 			c.JSON(http.StatusInternalServerError, gin.H{
 				"success": false,
